add more route

+++ src/Components/Utils/Localstorage.js
@@ -1,4 +1,5 @@
 
+
   
 export const saveToLocalData =(data)=>{
     const saveData =JSON.parse(localStorage.getItem("read"))|| [];
@@ -15,22 +16,30 @@ export const saveToLocalData =(data)=>{
 };
 export const getDataFromLocal =()=>{
     const getDefaultData= JSON.parse(localStorage.getItem("read"))|| [];
+    
+
     return  getDefaultData;
+    
 };
 
 export const saveLocalStorageData=(data)=>{
     const saveLocalstore = JSON.parse(localStorage.getItem("wishlist")) || [];
     const AlreadyExisted= saveLocalstore.find(i =>i.bookId===data.bookId);
-    if(!AlreadyExisted){
+    const saveData= JSON.parse(localStorage.getItem('read'))||[];
+
+    const  Existed =saveData.find(item =>item.bookId===data.bookId)
+    if(!AlreadyExisted && !Existed){
         saveLocalstore.push(data);
         localStorage.setItem("wishlist",JSON.stringify(saveLocalstore))
         alert('Add to data wishlist successfuly');
     }
     else{
-        alert('Already data Added');
+        alert('Already data Added to read');
     }
 };
 export const getDataLocalStorage=()=>{
     const getDefauldata=JSON.parse(localStorage.getItem("wishlist"))|| [];
     return getDefauldata;
 }
+
+
